Changed date time stamp
Changed the date time stamp from milliseconds since epoch time start to
something more useful i.e. 10/23/2015 9:10 PM for example.

+++ GBD/Header.cs
@@ -8,7 +8,7 @@
         public int serviceVersion { get; set; }
         public string serviceProvider { get; set; }
         public string serviceConsumer { get; set; }
-        public long messageDateTimeStamp { get; set; }
+        public double messageDateTimeStamp { get; set; }
         public string sendRuntime { get; set; }
         public string classification { get; set; }
     }
